Change vertex position type from vec3 to uvec3 in vertex shaders and update mesh layout to use PushInt for consistency

+++ res/shaders/block.vert
@@ -1,6 +1,6 @@
 #version 460 core
 
-layout(location = 0) in vec3 position; // Vertex position in local chunk space
+layout(location = 0) in uvec3 position; // Vertex position in local chunk space
 layout(location = 1) in uvec2 texIndex; // Texture column and row for atlas mapping (normalized)
 layout(location = 2) in uint face; // Face index (0-5 for 6 faces)
